app/requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
coolsms-python-sdk==2.0.3
decorator==4.2.1
defusedxml==0.5.0
Django==2.0.2
django-cors-headers==2.2.0
django-extensions==1.9.9
djangorestframework==3.7.7
djangorestframework-xml==1.3.0
djangorestframework-yaml==1.0.3
entrypoints==0.2.3
future==0.16.0
html5lib==1.0.1
httpie==0.9.9
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==4.1.1
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
pandocfilters==1.4.2
parso==0.1.1
permission==0.4.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.0
python-magic==0.4.15
pytz==2017.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
Send2Trash==1.5.0
serializers==0.2.4
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
views==0.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
